boutique/requirements.txt
asgiref==3.5.2
autopep8==1.7.0
certifi==2022.9.14
charset-normalizer==2.1.1
Django==3.2.5
djangorestframework==3.12.4
idna==3.4
pycodestyle==2.9.1
pytz==2022.2.1
requests==2.28.1
sqlparse==0.4.2
toml==0.10.2
urllib3==1.26.12
djangorestframework-simplejwt==4.7.2